LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-copyTo.any-expected.txt


PASS Test closed frame.
FAIL Test copying I420 frame to a non-shared ArrayBuffer assert_equals: allocationSize() expected 12 but got 0
FAIL Test copying I420 frame to a non-shared ArrayBufferView assert_equals: allocationSize() expected 12 but got 0
FAIL Test RGBA frame. assert_equals: allocationSize() expected 16 but got 0
PASS Test undersized buffer.
FAIL Test incorrect plane count. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
FAIL Test stride and offset work. assert_equals: allocationSize() expected 12 but got 0
FAIL Test stride and offset with padding. assert_equals: allocationSize() expected 19 but got 0
FAIL Test invalid stride. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
FAIL Test address overflow. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
FAIL Test codedRect. assert_equals: allocationSize() expected 12 but got 0
FAIL Test empty rect. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
FAIL Test unaligned rect. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
FAIL Test left crop. assert_equals: allocationSize() expected 6 but got 0
FAIL Test invalid rect. assert_throws_js: function "() => frame.allocationSize(options)" did not throw
